tafritio vako

+++ Cheque.cs
@@ -5,10 +5,10 @@
         public int ChequeNumber { get; set; }
         public int BankNumber { get; set; }
         public int SnifNumber { get; set; }
-        public int AccountNumber { get; set; }
+        public long AccountNumber { get; set; }
         public DateTime Date { get; set; }
         public decimal Sum { get; set; }
-        public Cheque(int chequeNumber, int bankNumber, int snifNumber, int accountNumber, DateTime date, decimal sum)
+        public Cheque(int chequeNumber, int bankNumber, int snifNumber, long accountNumber, DateTime date, decimal sum)
         {
             ChequeNumber = chequeNumber;
             BankNumber = bankNumber;
